requirements.txt
torch
torchvision
transformers
Pillow
numpy
wandb
matplotlib
nltk
torchtext